backend/requirements.txt
fastapi 
uvicorn 
pyjwt[crypto] 
passlib[bcrypt] 
sqlalchemy 
python-dotenv
pyodbc
pydantic[email]
fastapi-mail
authlib 
requests
python-multipart
httpx
itsdangerous
duckdb
pandas
ldap3
cryptography
scipy
scikit-learn
graphviz